feat(core): médaillon de résultat & bande KPI compacte promus en design-system
- EteeloResultMedallion (processing/succès/échec, halo, reduced-motion) extrait
  de l'encaissement vers core/widgets ; Facturation repointée dessus.
- EteeloKpiStrip (tuiles compactes label+valeur, liseré, responsive) extraite du
  détail Facturation vers core/components ; FinanceDetailKpiStrip devient un alias.

Réutilisés par la refonte de la Composition des classes, sans dépendance
classes -> finance.
